refactor design database code.
MasterDB is now a concrete class, and it knows how to create all
types of outputs.  TemplateDB and ModuleDB are now just convenience
subclasses of MasterDB that adds convenience attributes and function
aliases.

The reason for this refactoring is because model generation requires a
different approach to content list generation.  Also the code is
slightly more maintainable this way.